+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 - Reformat JS files. (#278, #281)
 - Add tests around callback feature verifying interests of actors are guaranteed. (#282)
 - Add IexecEscrow.v8 tests and developer notices. (#283)
+- Add PoCo Boost modules to a timelock controlled proxy. (#284)
 
 ## v5.4.2
 - Use latest Nethermind base image `nethermindeth/nethermind:iexec`